# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dunder Mifflin Delivery API",
        "description": "API documentation for the Dunder Mifflin delivery service",
        "contact": {
            "email": "support@dundermifflin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Log in to get the authentication token",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "your-access-token"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/museum": {
            "get": {
                "summary": "Search for Nearest Museums",
                "description": "List museums by proximity in a given municipality",
                "operationId": "37c2c67440b752f27aae309c8448cdce",
                "parameters": [
                    {
                        "name": "m",
                        "in": "query",
                        "description": "Municipality",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Latitude",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "long",
                        "in": "query",
                        "description": "Longitude",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "r",
                        "in": "query",
                        "description": "Radius (optional)",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Museum"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Museum"
                ],
                "summary": "Create Museum",
                "description": "Register a new museum. Requires user authentication.",
                "operationId": "storeMuseum",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "lat",
                                    "long",
                                    "region",
                                    "province",
                                    "municipality"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Museo nazionale di Capodimonte"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 40.9458662999999
                                    },
                                    "long": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 14.3715925
                                    },
                                    "region": {
                                        "type": "string",
                                        "example": "Campania"
                                    },
                                    "province": {
                                        "type": "string",
                                        "example": "NA"
                                    },
                                    "municipality": {
                                        "type": "string",
                                        "example": "Napoli"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "resource": {
                                            "$ref": "#/components/schemas/Museum"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Museum": {
                "title": "Museum",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Museo nazionale di Capodimonte"
                    },
                    "lat": {
                        "type": "number",
                        "format": "float",
                        "example": 40.9458663
                    },
                    "long": {
                        "type": "number",
                        "format": "float",
                        "example": 14.3715925
                    },
                    "region": {
                        "type": "string",
                        "example": "Campania"
                    },
                    "province": {
                        "type": "string",
                        "example": "NA"
                    },
                    "municipality": {
                        "type": "string",
                        "example": "Napoli"
                    },
                    "distance": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Museum",
            "description": "Museum"
        }
    ]
}